Minor updates (#227)
* Separate analytics class from resouce.py

* Add some missing endpoints

* bump version
