+++ twitter_ads/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (C) 2015 Twitter, Inc.
 
-VERSION = (6, 0, 0)
+VERSION = (6, 0, 1)
 API_VERSION = '6'
 
 from twitter_ads.utils import get_version
