+++ twitter_ads/creative.py
@@ -7,7 +7,8 @@ from twitter_ads import API_VERSION
 from twitter_ads.cursor import Cursor
 from twitter_ads.enum import TRANSFORM
 from twitter_ads.http import Request
-from twitter_ads.resource import resource_property, Resource, Persistence, Analytics
+from twitter_ads.analytics import Analytics
+from twitter_ads.resource import resource_property, Resource, Persistence
 from twitter_ads.utils import Deprecated, FlattenParams
